Delete RoleRequest - delete connected WF - fix

+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/model/service/impl/DefaultIdmRoleRequestService.java
@@ -3,6 +3,7 @@ package eu.bcvsolutions.idm.core.model.service.impl;
 import java.io.Serializable;
 import java.text.MessageFormat;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -51,6 +52,7 @@ import eu.bcvsolutions.idm.core.model.service.api.IdmIdentityService;
 import eu.bcvsolutions.idm.core.model.service.api.IdmRoleRequestService;
 import eu.bcvsolutions.idm.core.security.api.domain.BasePermission;
 import eu.bcvsolutions.idm.core.security.api.service.SecurityService;
+import eu.bcvsolutions.idm.core.workflow.model.dto.WorkflowFilterDto;
 import eu.bcvsolutions.idm.core.workflow.model.dto.WorkflowProcessInstanceDto;
 import eu.bcvsolutions.idm.core.workflow.service.WorkflowProcessInstanceService;
 
@@ -237,6 +239,9 @@ public class DefaultIdmRoleRequestService
 		Assert.notNull(requestId, "Role request ID is required!");
 		IdmRoleRequestDto request = this.getDto(requestId);
 		Assert.notNull(request, "Role request is required!");
+		if(RoleRequestState.APPROVED != request.getState()){
+			return request;
+		}
 
 		List<IdmConceptRoleRequestDto> concepts = request.getConceptRoles();
 		IdmIdentity identity = identityService.get(request.getApplicant());
@@ -453,6 +458,19 @@ public class DefaultIdmRoleRequestService
 
 	private void cancelWF(IdmRoleRequestDto dto) {
 		if (!Strings.isNullOrEmpty(dto.getWfProcessId())) {
+			WorkflowFilterDto filter = new WorkflowFilterDto();
+			filter.setProcessInstanceId(dto.getWfProcessId());
+			
+			Collection<WorkflowProcessInstanceDto> resources = workflowProcessInstanceService.searchInternal(filter, false).getResources();
+			if(resources.isEmpty()){
+				// Process with this ID not exist ... maybe was ended 
+				this.addToLog(dto,
+						MessageFormat.format(
+								"Workflow process with ID [{0}] was not deleted, because was not found. Maybe was ended before.",
+								dto.getWfProcessId()));
+				return;
+			}
+			
 			workflowProcessInstanceService.delete(dto.getWfProcessId(),
 					"Role request use this WF, was deleted. This WF was deleted too.");
 			this.addToLog(dto,
